ai-pdf-server/requirements.txt
# Python dependencies
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# PDF Processing
pymupdf>=1.23.0

# AI/ML
openai>=1.12.0
faiss-cpu>=1.7.4
numpy>=1.26.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1

# Logging
structlog>=24.1.0
